+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Publishing to YAML format by default
 - AsciiDoc custom role `scenario` added to cells in scenario column
 - AsciiDoc custom role `expectation` added to cells in columns where header name ends in `?`
 - Rendering explicit whitespace in AsciiDoc tables